+++ src/PlayGen.SUGAR.WebAPI/Controllers/AchievementsController.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+using PlayGen.SUGAR.Authorization;
+using PlayGen.SUGAR.Common.Shared.Permissions;
 using PlayGen.SUGAR.Contracts.Shared;
 using PlayGen.SUGAR.Core.Controllers;
 using PlayGen.SUGAR.Core.Utilities;
@@ -12,15 +16,17 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 	/// Web Controller that facilitates Achievement specific operations.
 	/// </summary>
 	[Route("api/[controller]")]
-	[Authorization]
 	public class AchievementsController : Controller
 	{
-		private readonly EvaluationController _evaluationCoreController;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly EvaluationController _evaluationCoreController;
 
-		public AchievementsController(EvaluationController evaluationController)
+		public AchievementsController(EvaluationController evaluationController,
+                    IAuthorizationService authorizationService)
 		{
-            _evaluationCoreController = _evaluationCoreController;
-		}
+            _evaluationCoreController = evaluationController;
+            _authorizationService = authorizationService;
+        }
 
 		/// <summary>
 		/// Find an Achievement that matches <param name="token"/> and <param name="gameId"/>.
@@ -32,12 +38,17 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		/// <returns>Returns <see cref="EvaluationResponse"/> that holds Achievement details</returns>
 		[HttpGet("find/{token}/{gameId:int}")]
 		[HttpGet("find/{token}/global")]
-		//[ResponseType(typeof(EvaluationResponse))]
-		public IActionResult Get([FromRoute]string token, [FromRoute]int? gameId)
+        //[ResponseType(typeof(EvaluationResponse))]
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Get, AuthorizationOperation.Achievement)]
+        public IActionResult Get([FromRoute]string token, [FromRoute]int? gameId)
 		{
-			var achievement = _evaluationCoreController.Get(token, gameId);
-			var achievementContract = achievement.ToContract();
-			return new ObjectResult(achievementContract);
+            if (_authorizationService.AuthorizeAsync(User, gameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                var achievement = _evaluationCoreController.Get(token, gameId);
+                var achievementContract = achievement.ToContract();
+                return new ObjectResult(achievementContract);
+            }
+            return Unauthorized();
 		}
 
 		/// <summary>
@@ -50,13 +61,18 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		/// <returns>Returns multiple <see cref="EvaluationResponse"/> that hold Achievement details</returns>
 		[HttpGet("global/list")]
 		[HttpGet("game/{gameId:int}/list")]
-		//[ResponseType(typeof(IEnumerable<EvaluationResponse>))]
-		public IActionResult Get([FromRoute]int? gameId)
+        //[ResponseType(typeof(IEnumerable<EvaluationResponse>))]
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Get, AuthorizationOperation.Achievement)]
+        public IActionResult Get([FromRoute]int? gameId)
 		{
-			var achievement = _evaluationCoreController.GetByGame(gameId);
-			var achievementContract = achievement.ToContractList();
-			return new ObjectResult(achievementContract);
-		}
+            if (_authorizationService.AuthorizeAsync(User, gameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                var achievement = _evaluationCoreController.GetByGame(gameId);
+                var achievementContract = achievement.ToContractList();
+                return new ObjectResult(achievementContract);
+            }
+            return Unauthorized();
+        }
 
 		/// <summary>
 		/// Find the current progress for all achievements for a <param name="gameId"/> for <param name="actorId"/>.
@@ -114,13 +130,18 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		[HttpPost("create")]
 		//[ResponseType(typeof(EvaluationResponse))]
 		[ArgumentsNotNull]
-		public IActionResult Create([FromBody] EvaluationCreateRequest newAchievement)
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Create, AuthorizationOperation.Achievement)]
+        public IActionResult Create([FromBody] EvaluationCreateRequest newAchievement)
 		{
-			var achievement = newAchievement.ToAchievementModel();
-            achievement = (Achievement)_evaluationCoreController.Create(achievement);
-			var achievementContract = achievement.ToContract();
-			return new ObjectResult(achievementContract);
-		}
+            if (_authorizationService.AuthorizeAsync(User, newAchievement.GameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                var achievement = newAchievement.ToAchievementModel();
+                achievement = (Achievement)_evaluationCoreController.Create(achievement);
+                var achievementContract = achievement.ToContract();
+                return new ObjectResult(achievementContract);
+            }
+            return Unauthorized();
+        }
 
 		/// <summary>
 		/// Update an existing Achievement.
@@ -130,10 +151,14 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		/// <param name="achievement"><see cref="EvaluationRequest"/> object that holds the details of the Achievement.</param>
 		[HttpPut("update")]
 		[ArgumentsNotNull]
-		public void Update([FromBody] EvaluationUpdateRequest achievement)
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Update, AuthorizationOperation.Achievement)]
+        public void Update([FromBody] EvaluationUpdateRequest achievement)
 		{
-			var achievementModel = achievement.ToAchievementModel();
-            _evaluationCoreController.Update(achievementModel);
+            if (_authorizationService.AuthorizeAsync(User, achievement.GameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                var achievementModel = achievement.ToAchievementModel();
+                _evaluationCoreController.Update(achievementModel);
+            }
 		}
 
 		/// <summary>
@@ -145,10 +170,13 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		/// <param name="gameId">ID of the Game the Achievement is for</param>
 		[HttpDelete("{token}/global")]
 		[HttpDelete("{token}/{gameId:int}")]
-		public void Delete([FromRoute]string token, [FromRoute]int? gameId)
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Delete, AuthorizationOperation.Achievement)]
+        public void Delete([FromRoute]string token, [FromRoute]int? gameId)
 		{
-            _evaluationCoreController.Delete(token, gameId);
+            if (_authorizationService.AuthorizeAsync(User, gameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                _evaluationCoreController.Delete(token, gameId);
+            }
 		}
-
 	}
 }
