+++ src/PlayGen.SUGAR.WebAPI/Controllers/LeaderboardsController.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+using PlayGen.SUGAR.Authorization;
+using PlayGen.SUGAR.Common.Shared.Permissions;
 using PlayGen.SUGAR.Contracts;
 using PlayGen.SUGAR.Contracts.Shared;
 using PlayGen.SUGAR.Core;
@@ -12,18 +16,20 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 	/// Web Controller that facilitates Leaderboard specific operations.
 	/// </summary>
 	[Route("api/[controller]")]
-	[Authorization]
 	public class LeaderboardsController : Controller
 	{
-		private readonly Data.EntityFramework.Controllers.LeaderboardController _leaderboardController;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly Data.EntityFramework.Controllers.LeaderboardController _leaderboardController;
 		private readonly LeaderboardController _leaderboardEvaluationController;
 
 		public LeaderboardsController(Data.EntityFramework.Controllers.LeaderboardController leaderboardController,
-			LeaderboardController leaderboardEvaluationController)
+			LeaderboardController leaderboardEvaluationController,
+            IAuthorizationService authorizationService)
 		{
 			_leaderboardController = leaderboardController;
 			_leaderboardEvaluationController = leaderboardEvaluationController;
-		}
+            _authorizationService = authorizationService;
+        }
 
 		/// <summary>
 		/// Find a list of leaderboards that match <param name="gameId"/>.
@@ -72,13 +78,18 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		[HttpPost("create")]
 		//[ResponseType(typeof(LeaderboardResponse))]
 		[ArgumentsNotNull]
-		public IActionResult Create([FromBody] LeaderboardRequest newLeaderboard)
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Create, AuthorizationOperation.Leaderboard)]
+        public IActionResult Create([FromBody] LeaderboardRequest newLeaderboard)
 		{
-			var leaderboard = newLeaderboard.ToModel();
-			_leaderboardController.Create(leaderboard);
-			var leaderboardContract = leaderboard.ToContract();
-			return new ObjectResult(leaderboardContract);
-		}
+            if (_authorizationService.AuthorizeAsync(User, newLeaderboard.GameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                var leaderboard = newLeaderboard.ToModel();
+                _leaderboardController.Create(leaderboard);
+                var leaderboardContract = leaderboard.ToContract();
+                return new ObjectResult(leaderboardContract);
+            }
+            return Unauthorized();
+        }
 
 		/// <summary>
 		/// Get the standings for a Leaderboard using a <see cref="LeaderboardStandingsRequest"/>.
@@ -104,10 +115,14 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		/// <param name="leaderboard"><see cref="LeaderboardRequest"/> object that holds the details of the Leaderboard.</param>
 		[HttpPut("update")]
 		[ArgumentsNotNull]
-		public void Update([FromBody] LeaderboardRequest leaderboard)
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Update, AuthorizationOperation.Leaderboard)]
+        public void Update([FromBody] LeaderboardRequest leaderboard)
 		{
-			var leaderboardModel = leaderboard.ToModel();
-			_leaderboardController.Update(leaderboardModel);
+            if (_authorizationService.AuthorizeAsync(User, leaderboard.GameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                var leaderboardModel = leaderboard.ToModel();
+                _leaderboardController.Update(leaderboardModel);
+            }
 		}
 
 		/// <summary>
@@ -119,9 +134,13 @@ namespace PlayGen.SUGAR.WebAPI.Controllers
 		/// <param name="gameId">ID of the Game the Leaderboard is for</param>
 		[HttpDelete("{token}/global")]
 		[HttpDelete("{token}/{gameId:int}")]
-		public void Delete([FromRoute]string token, [FromRoute]int? gameId)
+        [Authorization(ClaimScope.Game, AuthorizationOperation.Delete, AuthorizationOperation.Leaderboard)]
+        public void Delete([FromRoute]string token, [FromRoute]int? gameId)
 		{
-			_leaderboardController.Delete(token, gameId);
+            if (_authorizationService.AuthorizeAsync(User, gameId, (AuthorizationRequirement)HttpContext.Items["Requirements"]).Result)
+            {
+                _leaderboardController.Delete(token, gameId);
+            }
 		}
 	}
 }
